qwdqwd

+++ Matrix/Minor.c
@@ -5,18 +5,17 @@
 /*                                                    +:+ +:+         +:+     */
 /*   By: olahrizi <olahrizi@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2023/07/16 04:42:23 by aaitouna          #+#    #+#             */
-/*   Updated: 2023/11/15 04:01:56 by olahrizi         ###   ########.fr       */
+/*   Created: 2023/12/25 12:34:38 by olahrizi          #+#    #+#             */
+/*   Updated: 2023/12/25 12:35:42 by olahrizi         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "../header.h"
 
-#include "../minirt.h"
-
-void copy_clmn(double *clm1, double *clm2, int column, int size)
+void	copy_clmn(float *clm1, float *clm2, int column, int size)
 {
-	int j;
-	int fci;
+	int	j;
+	int	fci;
 
 	j = 0;
 	fci = 0;
@@ -31,11 +30,11 @@ void copy_clmn(double *clm1, double *clm2, int column, int size)
 	}
 }
 
-t_matrix *submatrix(t_matrix *mt, int row, int column)
+t_matrix	*submatrix(t_matrix *mt, int row, int column)
 {
-	int i;
-	int fri;
-	t_matrix *new_matrix;
+	int			i;
+	int			fri;
+	t_matrix	*new_matrix;
 
 	if (mt->rows <= 1 || mt->cols <= 1)
 		return (mt);
@@ -47,7 +46,7 @@ t_matrix *submatrix(t_matrix *mt, int row, int column)
 		if (i == row)
 		{
 			i++;
-			continue;
+			continue ;
 		}
 		copy_clmn(new_matrix->matrix[fri], mt->matrix[i], column, mt->cols);
 		fri++;
@@ -56,10 +55,10 @@ t_matrix *submatrix(t_matrix *mt, int row, int column)
 	return (new_matrix);
 }
 
-double minor(t_matrix *mt, int row, int column)
+float	minor(t_matrix *mt, int row, int column)
 {
-	t_matrix *sub;
-	double m_determinant;
+	t_matrix	*sub;
+	float		m_determinant;
 
 	sub = submatrix(mt, row, column);
 	m_determinant = determinant(sub);
